test(codex): split codegraph worker coverage
